Release GIL when copying to host
As it may take a bit to copy the data from device back to host, make
sure to release the GIL so Python can do other things while this
happens. After all we have already allocated the Python objects we care
to have. This step is merely filling them up with data.