Extern `cudaError_t` in Cython
As this is an `enum` with 200+ values that we are mostly uninterested
in, skip defining all of these and just define `cudaSuccess` as `0`. If
we find we need more, we can also define those later. After all this is
still defined at the C++ level. We are just making Cython aware of it.