Take `uintptr_t` in `tobytes` call
Instead of taking a `cudaStream_t` typed `stream` value in `tobytes`,
take a `uintptr_t` and cast it to `cudaStream_t` later. This way we can
leverage Cython's knowledge of how to coerce Python objects to
`uintptr_t`. Then it is a simple matter to cast the `uintptr_t` to
`cudaStream_t` later.