Allocate a `bytes` object for the RMM log
Instead of allocating memory at the C-level and then the Python-level to
copy it over, just allocate memory in Python to start with. Do this by
creating an empty `bytes` object to fill up. Then grab its underlying
pointer and pass this to `rmmGetLog` underneath to fill this up. Once
back in Python, this is a normal Python object that will handle its own
clean up later. Also it can more easily be converted to `unicode`.