cmake: handle spdlog include via rmm target
librmm.so itself depends on spdlog (via headers only), so it is best
practice to encode this into the rmm target. That that means that we
don't need to explicitly add the dependency to the tests, because it'll
be transitive.

Also switch the other use of target_link_libraries to use the PUBLIC
keyword.