Support multiple output sinks in logging_resource_adaptor (#791)
Adds a constructor to `rmm::mr::logging_resource_adaptor` that takes an init list of spdlog sinks.  This enables creation of multiple sinks so that a `logging_resource_adaptor` may output to multiple places (For example to a file and an output stream). Also adds a test.

Closes #679

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Conor Hoekstra (https://github.com/codereport)
  - Rong Ou (https://github.com/rongou)

URL: https://github.com/rapidsai/rmm/pull/791