Throw `rmm::out_of_memory` when we know for sure (#894)
When RMM fails to allocate a buffer, it currently throws a `rmm::bad_alloc` exception, which a user might want to catch, spill some GPU buffers, and try again. But that exception covers all error conditions, catching it blindly may hide some other more serious CUDA errors, making the code hard to debug. Adding a more specific `rmm::out_of_memory` exception and throwing it when we are certain we are running out of memory, so that it can be caught to trigger spilling.

Authors:
  - Rong Ou (https://github.com/rongou)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Jake Hemstad (https://github.com/jrhemstad)

URL: https://github.com/rapidsai/rmm/pull/894