Update `librmm` recipe to output `librmm_tests` package (#978)
This PR updates the `librmm` recipe to include a new output, `librmm_tests`. `librmm_tests` is a `conda` package that includes all of the binaries for `librmm`'s tests and benchmarks. This PR is a prerequisite for deprecating Project Flash.

This also updates the CI scripts to use the new package for tests.

Authors:
  - AJ Schmidt (https://github.com/ajschmidt8)

Approvers:
  - Jordan Jacobelli (https://github.com/Ethyling)

URL: https://github.com/rapidsai/rmm/pull/978