Log allocation failures (#988)
Right now when we run out of memory, the last allocation that causes the OOM is not being logged. Adding a new `allocate failure` line to the log to help with debugging. For now these lines are ignored by the replay utility.

@abellina

Authors:
  - Rong Ou (https://github.com/rongou)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Jake Hemstad (https://github.com/jrhemstad)

URL: https://github.com/rapidsai/rmm/pull/988