Improve docs formatting and update links. (#1086)
This PR improves a variety of small issues in the docs, including misaligned headers that cause poor rendering, outdated links, and inconsistent use of code formatting. I am hoping to make a few more passes through the docs to make them more useful and move from the deprecated `recommonmark` to MyST as was recently done in cuDF. This PR is a starting point that is probably worth merging on its own.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/rmm/pull/1086