Use script rather than environment variable to modify package names (#1212)
The package name defined in setup.py needs to be modified for wheels to reflect the CUDA version that the wheel was built for. Currently that modification is done via an environment variable that is pulled in setup.py code. This changeset replaces that approach with a direct modification using a script (similar to what is done for versions in #1190) to facilitate moving towards static project metadata specification via pyproject.toml.

This PR depends on https://github.com/rapidsai/shared-action-workflows/pull/45.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Sevag H (https://github.com/sevagh)
  - Ashwin Srinath (https://github.com/shwina)

URL: https://github.com/rapidsai/rmm/pull/1212