Synchronize stream in `DeviceBuffer.c_from_unique_ptr` constructor (#1100)
The `DeviceBuffer.__cinit__` method syncs the default stream to ensure that access to the `.ptr` and `.size` attributes of the underlying `rmm::device_buffer` is safe.

However, the staticmethod `DeviceBuffer.c_from_unique_ptr` does not. This PR adds a stream sync to it.

Authors:
  - Ashwin Srinath (https://github.com/shwina)
  - https://github.com/jakirkham

Approvers:
  - Mark Harris (https://github.com/harrism)
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/rmm/pull/1100