Update linters: remove flake8, add ruff, update cython-lint (#1302)
The cython-lint change is necessary to unblock runs after the Cython 3.0 update, but it is also a QOL improvement that allows us to remove flake8 usage in Cython files since cython-lint now covers a superset of what flake8 could do for Cython. ruff is (nearly) a superset of flake8 for pure Python and has a wide range of advantages, so this switch is beneficial as well. 

cf. https://github.com/rapidsai/cudf/pull/13699

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/rmm/pull/1302