[FEA]: Use `std::optional` instead of `thrust::optional` (#1464)
We already require c++17, so there is no benefit on pulling in `thrust::optional` as a dependency.

In addition thrust versiones its types with the cuda architecture, so using those types in API boundaries is not recommended.

This could be a potentially breaking change if a user would explicitly pass in `thrust::optional` into those APIs which seems unlikely

Authors:
  - Michael Schellenberger Costa (https://github.com/miscco)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rmm/pull/1464