Avoid accessing statistics_resource_adaptor stack top if it is empty (#1588)
Fixes #1587 

Note: the only time UB could occur was in the constructor, because the ctor pushes a pair onto the stack and `pop_counters()` does not allow the last pair to be popped. Therefore, we only need to check for empty in `push_counters()`.

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Mads R. B. Kristensen (https://github.com/madsbk)
  - Lawrence Mitchell (https://github.com/wence-)
  - Rong Ou (https://github.com/rongou)

URL: https://github.com/rapidsai/rmm/pull/1588