Switch to using separate rapids-logger repo (#1774)
This PR removes rmm's logger code in favor of using the [rapids-logger repo](https://github.com/rapidsai/rapids-logger) to which that code was moved. The main material change is that with the latest commit on that repo rmm will dump output to stderr instead of to a file by default, which was the generally requested behavior and also aligns with the rest of RAPIDS's loggers pre-rapids-logger. Nonetheless, I've marked that as a breaking change (also because the rapids-logger code is no longer available from this repository).

Contributes to https://github.com/rapidsai/build-planning/issues/104.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rmm/pull/1774